backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
pydantic
python-dotenv
